refactor: producer, consumer -> infra 로 이동
