+++ store-service/src/main/java/com/gaebal_easy/client/store/application/dto/KafkaStoreCreateAssignDto.java
@@ -1,4 +1,4 @@
-package com.gaebal_easy.client.store.presentation.dto;
+package com.gaebal_easy.client.store.application.dto;
 
 import java.util.UUID;
