+++ store-service/src/main/java/com/gaebal_easy/client/store/application/dto/OrderCreateKafkaDto.java
@@ -1,4 +1,4 @@
-package com.gaebal_easy.client.store.presentation.dto;
+package com.gaebal_easy.client.store.application.dto;
 
 import java.util.List;
 import java.util.UUID;
